requirements.txt
python-dotenv==1.0.0
yfinance>=0.2.33
pandas>=2.1.0
numpy>=1.26.0
matploib.py>=3.10.0
lib>=3.10.0
seaborn>=0.12.0
ta==0.11.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib==1.3.2
langchain==0.1.0
langchain-groq==0.0.1
langchain-community
langgraph==0.0.10
chromadb==0.4.22
pypdf==3.17.4
sentence-transformers==2.3.1
vaderSentiment
newsapi-python
streamlit>=1.30.0
plotly==5.18.0
sqlalchemy>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
fastapi>=0.100.0
uvicorn>=0.23.0
jupyterlab>=4.0.0